refactor(widgets/errors): refactor app error widgets.
