+++ lib/interfaces/router/routes/res/movies_route.dart
@@ -21,21 +21,23 @@ Widget _movieDetailBuilder(BuildContext context, GoRouterState router) {
     return MovieDetailView(movieId);
   } catch (e) {
     if (e is FormatException) {
-      return BadRequestErrorView(
-        message: 'Invalid path format. '
-            'Valid movie id are numbers '
-            '(ex: "/movie-123").',
-        action: ElevatedButton.icon(
-          onPressed: () => context.pop(),
-          icon: const Icon(Icons.arrow_back),
-          label: const Text('Go Back'),
+      return Scaffold(
+        body: AppHttpErrorWidget(
+          statusCode: 400,
+          message: 'Invalid path format. '
+              'Valid movie id are numbers (ex: "/movie-123").',
+          action: ElevatedButton.icon(
+            onPressed: () => context.pop(),
+            icon: const Icon(Icons.arrow_back),
+            label: const Text('Go Back'),
+          ),
         ),
       );
     }
 
     kLogger.f('GoRouter (${router.uri.path}) builder', error: e);
 
-    return const UnknownErrorWidget();
+    return const Scaffold(body: AppErrorWidget());
   }
 }
 
@@ -49,20 +51,23 @@ Widget _movieCollectionDetailBuilder(
     return MovieCollectionDetailView(collectionId);
   } catch (e) {
     if (e is FormatException) {
-      return BadRequestErrorView(
-        message: 'Invalid path format. '
-            'Valid movie collection id are numbers '
-            '(ex: "/movie-collection-123").',
-        action: ElevatedButton.icon(
-          onPressed: () => context.pop(),
-          icon: const Icon(Icons.arrow_back),
-          label: const Text('Go Back'),
+      return Scaffold(
+        body: AppHttpErrorWidget(
+          statusCode: 400,
+          message: 'Invalid path format. '
+              'Valid movie collection id are numbers '
+              '(ex: "/movie-collection-123").',
+          action: ElevatedButton.icon(
+            onPressed: () => context.pop(),
+            icon: const Icon(Icons.arrow_back),
+            label: const Text('Go Back'),
+          ),
         ),
       );
     }
 
     kLogger.f('GoRouter (${router.uri.path}) builder', error: e);
 
-    return const UnknownErrorWidget();
+    return const Scaffold(body: AppErrorWidget());
   }
 }
