+++ lib/interfaces/utils/riverpod_async_value_handlers.dart
@@ -4,8 +4,8 @@ import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 
 import 'package:screen_pal/configs/constants.dart';
+import 'package:screen_pal/interfaces/widgets/errors/app_error_widget.dart';
 import 'package:screen_pal/interfaces/widgets/errors/dio_exception_widget.dart';
-import 'package:screen_pal/interfaces/widgets/errors/unknown_error_widget.dart';
 
 abstract final class RiverpodAsyncValueHandlers {
   /// [action] used to add refresh provider action.
@@ -30,7 +30,7 @@ abstract final class RiverpodAsyncValueHandlers {
 
     kLogger.f('RiverpodAsyncValueHandlers.error', error: error);
 
-    return UnknownErrorWidget(action: actionButton);
+    return AppErrorWidget(action: actionButton);
   }
 
   static Widget loading() => const Center(child: CircularProgressIndicator());
